requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
PyNuclino==0.2.1
python-dotenv==1.0.1
ratelimit==2.2.1
requests==2.32.3
urllib3==2.3.0
